requirements.txt
aiogram>=3.23.0,<4.0.0
aiohttp>=3.12.0,<4.0.0
python-dotenv>=1.0.1,<2.0.0